+++ package.json
@@ -98,6 +98,7 @@
   "workspaces": [
     "example"
   ],
+  "packageManager": "yarn@3.6.1",
   "jest": {
     "preset": "react-native",
     "modulePathIgnorePatterns": [
